Feature ETP-574: Fix build.
