+++ packages/ComponentLibrary/src/components/ProfileModal/ToggleSection/index.tsx
@@ -10,6 +10,7 @@ import { InputPassword, theme } from '../..';
 import LockOutlined from '../../../assets/icons/lock.svg';
 import Select from '../../Input/Select';
 import { useTranslation } from '../../../../../MainUI/hooks/useTranslation';
+import { Warehouse } from '@workspaceui/etendohookbinder/api/types';
 
 const icons: { [key in Item]: React.ReactElement } = {
   [Item.Role]: <></>,
@@ -43,7 +44,7 @@ const SelectorList: React.FC<SelectorListProps> = ({
   const warehouses = useMemo(() => {
     if (selectedRole) {
       const role = roles.find(r => r.id === selectedRole.value);
-      return role ? role.orgList.flatMap((org: { warehouseList: unknown; }) => org.warehouseList) : [];
+      return role ? role.orgList.flatMap((org: { warehouseList: Warehouse[] }) => org.warehouseList) : [];
     }
     return [];
   }, [roles, selectedRole]);
@@ -60,7 +61,7 @@ const SelectorList: React.FC<SelectorListProps> = ({
 
   const warehouseOptions = useMemo(
     () =>
-      warehouses.map((warehouse: { name: string; id: string; }) => ({
+      warehouses.map(warehouse => ({
         title: warehouse.name,
         value: warehouse.id,
         id: warehouse.id,
