+++ packages/ComponentLibrary/src/components/ProfileModal/types.ts
@@ -1,8 +1,8 @@
 import { ReactNode } from 'react';
 import { Section } from './ToggleButton/types';
 import { Item } from '../enums/index';
-import { Role } from '../../../../MainUI/src/contexts/types';
 import { Option } from '../Input/Select/types';
+import { Role } from '@workspaceui/etendohookbinder/api/types';
 
 export interface User {
   photoUrl: string;
